requirements.txt
alabaster==1.0.0
arrow==1.3.0
asgiref==3.8.1
babel==2.16.0
binaryornot==0.4.4
bleach==6.2.0
blinker==1.8.2
bokeh==3.6.0
certifi==2024.8.30
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
ConfigUpdater==3.2
contourpy==1.3.0
cookiecutter==2.6.0
cycler==0.12.1
dash==2.18.1
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
distlib==0.3.9
Django==4.2.16
docutils==0.21.2
filelock==3.16.1
Flask==3.0.3
fonttools==4.54.1
identify==2.6.1
idna==3.10
imagesize==1.4.1
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
linkify-it-py==2.0.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
myst-parser==4.0.0
nest-asyncio==1.6.0
nodeenv==1.9.1
numpy==2.1.2
packaging==24.1
pandas==2.2.3
panel==1.5.3
param==2.1.1
pillow==11.0.0
platformdirs==4.3.6
plotly==5.24.1
pre_commit==4.0.1
Pygments==2.18.0
pyparsing==3.2.0
PyScaffold==4.6
pyscaffoldext-cookiecutter==0.1
pyscaffoldext-custom-extension==0.6.3
pyscaffoldext-django==0.2
pyscaffoldext-dsproject==0.7.2
pyscaffoldext-markdown==0.5
pyscaffoldext-travis==0.3
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.2
pyviz_comms==3.0.3
PyYAML==6.0.2
requests==2.32.3
retrying==1.3.4
rich==13.9.3
setuptools-scm==8.1.0
six==1.16.0
snowballstemmer==2.2.0
Sphinx==8.1.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.1
tenacity==9.0.0
text-unidecode==1.3
tomli==2.0.2
tomlkit==0.13.2
tornado==6.4.1
tqdm==4.66.6
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
urllib3==2.2.3
virtualenv==20.27.1
webencodings==0.5.1
Werkzeug==3.0.6
xyzservices==2024.9.0
zipp==3.20.2
